Fix regression in Hive Generic UDF support on Databricks 12.2
