+++ sql-plugin/src/main/scala/org/apache/spark/sql/hive/rapids/rowBasedHiveUDFs.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, NVIDIA CORPORATION.
+ * Copyright (c) 2021-2023, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,15 +21,13 @@ import scala.collection.JavaConverters._
 import com.nvidia.spark.rapids.{GpuExpression, GpuLiteral, GpuRowBasedUserDefinedFunction, GpuScalar}
 import org.apache.hadoop.hive.ql.exec.{FunctionRegistry, UDF}
 import org.apache.hadoop.hive.ql.udf.{UDFType => HiveUDFType}
-import org.apache.hadoop.hive.ql.udf.generic.GenericUDF
-import org.apache.hadoop.hive.ql.udf.generic.GenericUDF.DeferredObject
 import org.apache.hadoop.hive.ql.udf.generic.GenericUDFUtils.ConversionHelper
-import org.apache.hadoop.hive.serde2.objectinspector.{ConstantObjectInspector, ObjectInspector, ObjectInspectorFactory}
+import org.apache.hadoop.hive.serde2.objectinspector.{ObjectInspector, ObjectInspectorFactory}
 import org.apache.hadoop.hive.serde2.objectinspector.ObjectInspectorFactory.ObjectInspectorOptions
 
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.{Expression, Literal, SpecializedGetters}
-import org.apache.spark.sql.hive.{DeferredObjectAdapter, HiveInspectors}
+import org.apache.spark.sql.hive.HiveInspectors
 import org.apache.spark.sql.hive.HiveShim.HiveFunctionWrapper
 import org.apache.spark.sql.types.DataType
 
@@ -147,44 +145,4 @@ case class GpuRowBasedHiveSimpleUDF(
   override def foldable: Boolean = udfDeterministic && children.forall(_.foldable)
 
   override def sql: String = s"$name(${children.map(_.sql).mkString(", ")})"
-}
-
-/** Row-based version of Spark's `HiveGenericUDF` running in a GPU operation */
-case class GpuRowBasedHiveGenericUDF(
-    name: String,
-    funcWrapper: HiveFunctionWrapper,
-    children: Seq[Expression]) extends GpuRowBasedHiveUDFBase {
-
-  @transient
-  override lazy val function: GenericUDF = funcWrapper.createFunction[GenericUDF]()
-
-  @transient
-  private lazy val returnInspector =
-    function.initializeAndFoldConstants(argumentInspectors.toArray)
-
-  @transient
-  private lazy val deferredObjects = argumentInspectors.zip(children).map {
-    case (inspect, child) => new DeferredObjectAdapter(inspect, child.dataType)
-  }.toArray
-
-  @transient
-  private lazy val unwrapper = unwrapperFor(returnInspector)
-
-  override protected def evaluateRow(childrenRow: InternalRow): Any = {
-    returnInspector // Make sure initialized.
-
-    var i = 0
-    val length = children.length
-    while (i < length) {
-      val idx = i
-      deferredObjects(i).set(() => childRowAccessors(idx)(childrenRow))
-      i += 1
-    }
-    unwrapper(function.evaluate(deferredObjects.asInstanceOf[Array[DeferredObject]]))
-  }
-
-  override lazy val dataType: DataType = inspectorToDataType(returnInspector)
-
-  override def foldable: Boolean =
-    udfDeterministic && returnInspector.isInstanceOf[ConstantObjectInspector]
 }
